Implicit gas tile & decal states (#15146)
